+++ lib/src/clean_module_creators/clean_module_i_creators.dart
@@ -1,0 +1,14 @@
+import 'dart:io';
+
+abstract class CleanModuleISSLCreator {
+  Future<void> create();
+}
+
+abstract class CleanModuleIDirectoryCreator {
+  Future<bool> createDirectories(); 
+  Directory get moduleDir;
+}
+
+abstract class CleanModuleIFileCreator {
+  Future<void> createNecessaryFiles();
+}