+++ CHANGELOG.md
@@ -419,3 +419,20 @@
 - Automatic Doc generation integrate
 ## 3.1.6
 - override command added for  config.json
+
+## 4.0.0
+- clean architecture integrate
+- Dependency injection integrate
+- default theme integrate
+- kts file and groovy wise build.gradle update
+- release and debug build wise build.gradle update
+- pubspec.yaml package update
+- clean architecture module integrate
+- Based on clean architecture Riverpod and Bloc pattern create
+- Asset file issue resolve
+- Riverpod and Bloc pattern module create
+- flavorType rename to mode
+
+
+
+